requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
psycopg2-binary
python-jose
passlib[bcrypt]
pandas
langgraph
reportlab
